rename security to bug-bounty
