+++ src/components/elements/off-canvas-menu.js
@@ -30,7 +30,7 @@ import Options from 'images/svg/custom/options-nav.svg'
 import Partner from 'images/svg/menu/partner.svg'
 import Payment from 'images/svg/menu/payment-methods.svg'
 import PaymentAgent from 'images/svg/menu/payment-agent.svg'
-import Security from 'images/svg/menu/security.svg'
+import BugBounty from 'images/svg/menu/bug-bounty.svg'
 import Regulatory from 'images/svg/menu/regulatory.svg'
 import SecureTrading from 'images/svg/menu/secure-trading.svg'
 import Smarttrader from 'images/svg/custom/smarttrader.svg'
@@ -532,11 +532,11 @@ export const OffCanvasMenuWrapper = (props) => {
                             </div>
                             <span>{localize('API')}</span>
                         </StyledLink>
-                        <StyledLink to="/security/" onClick={handleArrowClick}>
+                        <StyledLink to="/bugbounty/" onClick={handleArrowClick}>
                             <div>
-                                <img src={Security} alt="" width="32" height="32" />
+                                <img src={BugBounty} alt="" width="32" height="32" />
                             </div>
-                            <span>{localize('Security')}</span>
+                            <span>{localize('Bug Bounty')}</span>
                         </StyledLink>
                     </AccordionItem>
                 </Accordion>
@@ -603,11 +603,11 @@ export const OffCanvasMenuPartner = (props) => {
                     </div>
                     <span>{localize('API')}</span>
                 </StyledLink>
-                <StyledLink to="/security/" onClick={handleArrowClick}>
+                <StyledLink to="/bugbounty/" onClick={handleArrowClick}>
                     <div>
-                        <img src={Security} alt="" width="32" height="32" />
+                        <img src={BugBounty} alt="" width="32" height="32" />
                     </div>
-                    <span>{localize('Security')}</span>
+                    <span>{localize('Bug Bounty')}</span>
                 </StyledLink>
             </OffCanvasMenuContainer>
         </OffCanvasMenuSecondary>
