+++ src/components/layout/footer/main-links.js
@@ -161,7 +161,7 @@ const MainLinksSection = ({ is_ppc, is_ppc_redirect, is_eu_country }) => {
                             </Link>
                         </LinkWrapper>
                         <LinkWrapper>
-                            <Link to="/security/">{localize('Security')}</Link>
+                            <Link to="/bugbounty/">{localize('Bug Bounty')}</Link>
                         </LinkWrapper>
                     </LinksCol>
                     <LinksCol>
