+++ src/components/layout/nav.js
@@ -30,7 +30,7 @@ import {
 // Icons
 import Logo from 'images/svg/layout/logo-deriv.svg'
 import LogoPartner from 'images/svg/layout/logo-partners.svg'
-import LogoSecurity from 'images/svg/layout/logo-security.svg'
+import LogoBugBounty from 'images/svg/layout/logo-bug-bounty.svg'
 import Hamburger from 'images/svg/layout/hamburger_menu.svg'
 import Close from 'images/svg/layout/close-long.svg'
 import LogoOnly from 'images/svg/layout/logo-deriv-only.svg'
@@ -1046,8 +1046,8 @@ export const NavPartners = ({ no_login_signup }) => {
                                 </StyledLink>
                             </NavLink>
                             <NavLink>
-                                <StyledLink to="/security/" aria-label={localize('security')}>
-                                    {localize('Security')}
+                                <StyledLink to="/bugbounty/" aria-label={localize('Bug Bounty')}>
+                                    {localize('Bug Bounty')}
                                 </StyledLink>
                             </NavLink>
                         </StyledNavCenter>
@@ -1174,8 +1174,8 @@ export const NavSecurity = () => {
                 <StyledNavPartners>
                     <StyledNavWrapper>
                         <NavLeftPartners>
-                            <NavLogoLink to="/" aria-label={localize('Security')}>
-                                <img src={LogoSecurity} alt="logo security" />
+                            <NavLogoLink to="/" aria-label={localize('Bug Bounty')}>
+                                <img src={LogoBugBounty} alt="reslogo" />
                             </NavLogoLink>
                         </NavLeftPartners>
                         <SecurityNavRight button_ref={button_ref} mounted={true}>
@@ -1196,9 +1196,12 @@ export const NavSecurity = () => {
                                 <LogoLinkMobile
                                     style={{ margin: 'unset' }}
                                     to="/"
-                                    aria-label={localize('Security')}
+                                    aria-label={localize('Bug Bounty')}
                                 >
-                                    <SecurityLogoInMobile src={LogoSecurity} alt="logo security" />
+                                    <SecurityLogoInMobile
+                                        src={LogoBugBounty}
+                                        alt="logo bug bounty"
+                                    />
                                 </LogoLinkMobile>
 
                                 <LinkButton
