+++ src/images/common/bug-bounty/bug-report-desktop.png
