+++ src/images/common/bug-bounty/bug-report-mobile.png
