+++ src/pages/bugbounty/_about-deriv.tsx
@@ -8,7 +8,7 @@ import device from 'themes/device'
 
 const query = graphql`
     {
-        deriv_platform: file(relativePath: { eq: "security/devices.jpg" }) {
+        deriv_platform: file(relativePath: { eq: "bug-bounty/devices.jpg" }) {
             ...fadeIn
         }
     }
