+++ src/pages/bugbounty/_lazy-load.ts
