+++ src/pages/bugbounty/_submit_a_bug_report.tsx
@@ -57,10 +57,10 @@ const StyledHeader = styled(Header)`
 
 const query = graphql`
     query {
-        bug_report_desktop: file(relativePath: { eq: "security/bug-report-desktop.png" }) {
+        bug_report_desktop: file(relativePath: { eq: "bug-bounty/bug-report-desktop.png" }) {
             ...heroImage
         }
-        bug_report_mobile: file(relativePath: { eq: "security/bug-report-mobile.png" }) {
+        bug_report_mobile: file(relativePath: { eq: "bug-bounty/bug-report-mobile.png" }) {
             ...heroImage
         }
     }
