+++ src/pages/bugbounty/components/_hero_component.tsx
