+++ src/pages/bugbounty/static/_scope_content.ts
