+++ src/pages/bugbounty/vulnerabilities-types/index.tsx
